refactor spark
